% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sparse.R
\name{readsparse_nonascii_support}
\alias{readsparse_nonascii_support}
\title{Determine if readsparse will handle non-ASCII file paths}
\usage{
readsparse_nonascii_support()
}
\value{
A logical/boolean value telling whether non-ASCII file names will be
supported or not.
}
\description{
See \link{read.sparse} for details. In general, the library will
be able to handle non-ASCII file paths, unless it is compiled with G++ version
4 or earlier (the default under Windows for older versions of RTools, such as
RTools35).
}
